poc: minted reward 2
